requirements-osx.txt
paddleocr==2.9.1
paddlepaddle
pillow
pydantic
pandas
opencv-python
